% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\docType{data}
\name{coronavirus}
\alias{coronavirus}
\title{General Situation of corona virus in the World}
\format{
a tibble with 181335 observations and 7 variables:
\itemize{
\item \strong{date}: The date of the summary
\item \strong{province}: The province or state, when applicable
\item \strong{country}: The country or region name
\item \strong{lat}: Latitude point
\item \strong{long}: Longitude point
\item \strong{type}: the type of case (i.e., confirmed, death)
\item \strong{cases}: the number of daily cases (corresponding to the case type)
}
}
\source{
\href{https://www.who.int/}{WHO} and \href{https://github.com/RamiKrispin/coronavirus}{coronavirus}
}
\usage{
coronavirus
}
\description{
A dataset containing general information of corona virus about 188 countries. The
information includes daily confirmed cases, death and recovered cases.
}
\keyword{datasets}
