% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Easykable.R
\name{Easykable}
\alias{Easykable}
\title{Create a formatted and stylish table contains the Covid-19 information in the kableExtra function.}
\usage{
Easykable(n_country)
}
\arguments{
\item{n_country}{The number of the countries with most confirmed cases in the data to be formatted in table.}
}
\value{
A formatted table of the COVID data set with the specific obervations.
}
\description{
Creates a styled and formatted table to represent the summary case situation of COVID19 in each country.
}
