% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputselect.R
\name{Inputselect}
\alias{Inputselect}
\title{Run the shiny:selectInput for the Covid-19 app}
\usage{
Inputselect(inputID, chosen)
}
\arguments{
\item{inputID}{The variable which shows in the SelectInput function}

\item{chosen}{The information of the specific observation in that column which chose by the user}
}
\description{
This function can provide a more concise way to select desired input in the Shiny app
}
