% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shiny app to explore the further information about COVID-19}
\usage{
launch_app()
}
\description{
Launch the shiny app to explore the further information about COVID-19
}
