% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\docType{data}
\name{coronavirus}
\alias{coronavirus}
\title{General Situation of COVID-19 in the world}
\format{
An object of class \code{data.frame} with 181335 rows and 7 columns.
}
\source{
coronavirus
}
\usage{
coronavirus
}
\description{
This dataset contains daily information of COVID-19 about 188 countries.
}
\keyword{datasets}
