% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inputselect.R
\name{Inputselect}
\alias{Inputselect}
\title{Run the shiny:selectInput for the Covid-19 app}
\usage{
Inputselect(ID)
}
\arguments{
\item{ID}{The variable which shows in the SelectInput function}
}
\value{
A shiny control which can choose the specfic inputs for the following presentation in the Shiny app
}
\description{
This function can provide a more concise way to select desired input in the Shiny app
}
\examples{
\dontrun{
Inputselect("country")
} 

}
